Change type of all members of DIRCOUNT to LONG
This saves some code space by preventing unnecessary conversions
to/from WORD.

+++ desk/deskdir.h
@@ -14,8 +14,8 @@
  * typedefs
  */
 typedef struct {    /* #files/#folders/total filesize of directory & its subdirectories */
-    WORD files;
-    WORD dirs;
+    LONG files;
+    LONG dirs;
     LONG size;
 } DIRCOUNT;
